Rename embed.complexRecord to embed.recordWithMedia
