+++ lexicons/app/bsky/embed/recordWithMedia.json
@@ -1,6 +1,6 @@
 {
   "lexicon": 1,
-  "id": "app.bsky.embed.complexRecord",
+  "id": "app.bsky.embed.recordWithMedia",
   "description": "A representation of a record embedded in another form of content, alongside other compatible embeds",
   "defs": {
     "main": {
