+++ packages/api/src/client/types/app/bsky/feed/defs.ts
@@ -9,7 +9,7 @@ import * as AppBskyActorDefs from '../actor/defs'
 import * as AppBskyEmbedImages from '../embed/images'
 import * as AppBskyEmbedExternal from '../embed/external'
 import * as AppBskyEmbedRecord from '../embed/record'
-import * as AppBskyEmbedComplexRecord from '../embed/complexRecord'
+import * as AppBskyEmbedRecordWithMedia from '../embed/recordWithMedia'
 
 export interface PostView {
   uri: string
@@ -20,7 +20,7 @@ export interface PostView {
     | AppBskyEmbedImages.View
     | AppBskyEmbedExternal.View
     | AppBskyEmbedRecord.View
-    | AppBskyEmbedComplexRecord.View
+    | AppBskyEmbedRecordWithMedia.View
     | { $type: string; [k: string]: unknown }
   replyCount?: number
   repostCount?: number
